+++ lib/drag_and_drop_list_target.dart
@@ -1,18 +1,20 @@
+import 'package:drag_and_drop_lists/drag_and_drop_builder_parameters.dart';
 import 'package:drag_and_drop_lists/drag_and_drop_list_interface.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import 'package:drag_and_drop_lists/drag_and_drop_builder_parameters.dart';
 
 class DragAndDropListTarget extends StatefulWidget {
   final Widget child;
   final DragAndDropBuilderParameters parameters;
   final Function(DragAndDropListInterface newOrReordered,
       DragAndDropListTarget receiver) onDropOnLastTarget;
+  final double lastListTargetSize;
 
   DragAndDropListTarget(
       {this.child,
       @required this.parameters,
       @required this.onDropOnLastTarget,
+      this.lastListTargetSize = 110,
       Key key})
       : super(key: key);
 
@@ -45,8 +47,12 @@ class _DragAndDropListTarget extends State<DragAndDropListTarget>
         ),
         widget.child ??
             Container(
-              height: widget.parameters.axis == Axis.vertical ? 110 : null,
-              width: widget.parameters.axis == Axis.horizontal ? 110 : null,
+              height: widget.parameters.axis == Axis.vertical
+                  ? widget.lastListTargetSize
+                  : null,
+              width: widget.parameters.axis == Axis.horizontal
+                  ? 110
+                  : widget.lastListTargetSize,
             ),
       ],
     );