+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.1.1] - 28 September 2020
+
+* Added option for item dividers. See the `itemDivider` parameter in `DragAndDropLists`
+* Added option for inner list box decoration. See the `listInnerDecoration` parameter in `DragAndDropLists`
+* Added option for drag handles
+* Configurable bottom padding for list and items. See the `lastItemTargetHeight`, `addLastItemTargetHeightToTop` and `lastListTargetSize` parameters in `DragAndDropLists`
+
 ## [0.1.0] - 21 September 2020
 
 * Added canDrag option for lists